data:fix plmn id not properly displayed issue
If uint8_t type value and has a-f characters, std::cout will output
the represented character.
Explicitly cast a uint8_t variable to an int in order to get the
numerical representation.

CRs-Fixed: 3980072

+++ apps/tests/telsdk_console_app/Data/serving_system/ServingSystemListener.cpp
@@ -169,14 +169,16 @@ void ServingSystemListener::onLteAttachFailure(const telux::data::LteAttachFailu
         << ", Code: " << DataUtils::callEndReasonCode(info.rejectReason) << std::endl;
     std::cout << " PLMN:";
     for (unsigned int i = 0; i < info.plmnId.size(); ++i) {
-        std::cout << std::setfill('0') << std::setw(2) << std::hex << info.plmnId[i];
+        std::cout << std::setfill('0') << std::setw(2) << std::hex << static_cast<int>(
+            info.plmnId[i]);
     }
     std::cout << std::endl;
 
     if (info.primaryPlmnId.size()) {
         std::cout << " Primary PLMN:";
         for (unsigned  int i = 0; i < info.primaryPlmnId.size(); ++i) {
-            std::cout << std::setfill('0') << std::setw(2) << std::hex << info.primaryPlmnId[i];
+            std::cout << std::setfill('0') << std::setw(2) << std::hex << static_cast<int>(
+                info.primaryPlmnId[i]);
         }
         std::cout << std::endl;
     }
